refactor: pages and widgets
